% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{car_costs}
\alias{car_costs}
\title{Calculate total car travel cost}
\usage{
car_costs(START, STOP, travelers_number, fuel_consumption, fuel_price,
  insurence_fee, parking_time)
}
\arguments{
\item{START}{character. An address of starting point.}

\item{STOP}{character. An address of destination.}

\item{travelers_number}{numeric. Number of travelers. Default value is 1.}

\item{fuel_consumption}{numeric. Fuel consumption expressed as liters per 100 kilometers (L/100 km)}

\item{fuel_price}{numeric. Fuel price in PLN}

\item{insurence_fee}{numeric. Annual car insurance cost in PLN.}

\item{parking_time}{numeric. Parking time in hours.}
}
\value{
car_costs returns a cost of car travel per traveler in PLN
}
\description{
Calculate total car travel cost
}
\examples{
car_travel_cost(START="plac Defilad 1, 00-901 Warszawa Polska",
STOP="Stanisława Kostki Potockiego 10/16, Warszawa Polska",
fuel_consumption=8, fuel_price=4.5, travelers_number = 1,
insurence_fee=1000, parking_time=8)
}
