% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{car_travel_cost}
\alias{car_travel_cost}
\title{Calculate distans, amount of petrol, cost of petrol, cost of petrol for a given travel by car}
\usage{
car_travel_cost(START, STOP, combustion, fuel_price, travelers_number = 1)
}
\arguments{
\item{START}{character. An adress of starting point.}

\item{STOP}{character. An adresss of destination.}
}
\value{
car_travel_cost returns a list containing following components
}
\description{
Calculate distans, amount of petrol, cost of petrol, cost of petrol for a given travel by car
}
\examples{
car_travel_cost(START="plac Defilad 1, 00-901 Warszawa Polska",
STOP="Stanisława Kostki Potockiego 10/16, Warszawa Polska",
combustion=8, fuel_price=4.5, travelers_number = 1)
}
