% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_parking.R
\name{cost_of_parking}
\alias{cost_of_parking}
\title{Calculate cost for parking}
\usage{
cost_of_parking(parking_time)
}
\arguments{
\item{parking_time}{numeric. Parking time in hours.}
}
\value{
car_travel_cost returns daily cost of parking in the paid parking zone in Warsaw.
}
\description{
cost_of_parking calculates cost of parking in the paid parking zone in Warsaw.
}
\examples{
cost_of_parking(parking_time=3)
}
