% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztm_costs.R
\name{tickets_costs_annual_ZTM}
\alias{tickets_costs_annual_ZTM}
\title{Calculate minimal annual cost of optimal ticket}
\usage{
tickets_costs_annual_ZTM(travel_time, discount, zone, varsovian)
}
\arguments{
\item{travel_time}{numeric. Daily travel time in minutes.}

\item{discount}{ligical. Entitlement to a discount of 50\%.}

\item{zone}{numeric. Ticket zone 1, 2 or 3 (3 means zone 1 & 2)}

\item{varsovian}{logical. Entitlement to a Varsovian ticket discount}
}
\value{
tickets_costs_annual_ZTM returns a list containing following components:
  \item{ztm_ticket}{data.frame. Attributes of optimal ticket}
  \item{number_of_tickets}{Number of optimal tickets}
  \item{tickets_cost}{Minimal total annual cost of commuting for optimal tickets in PLN}
}
\description{
tickets_costs_annual_ZTM chooses optimat tickets and calculates minimal annual
cost of commuting by warsaw public transport.
}
\examples{
tickets_costs_annual_ZTM(travel_time=60, discount=TRUE, zone=1, varsovian=TRUE)
}
