% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ztm_price_table}
\alias{ztm_price_table}
\title{Prices of tickets applicable in vehicles
operated on the lines organised by ZTM.}
\format{A data frame with 34 rows and 7 variables:
\describe{
  \item{Lp}{ID number}
  \item{zone}{ticket zone}
  \item{time_min}{validity time in minutes}
  \item{discount}{the right to 50\% discount}
  \item{price}{standard price - fare}
  \item{Warszawiak}{the right to Varsovian ticket discount}
  \item{name}{Types of ticket}
  ...
}}
\usage{
ztm_price_table
}
\description{
A dataset containing the prices
and other attributes of 34 different types of ZTM tickets.
}
\keyword{datasets}
