% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{car_costs}
\alias{car_costs}
\title{Calculate total car travel cost}
\usage{
car_costs(START, STOP, travelers_number, combustion, fuel_price, insurence_fee,
  parking_time)
}
\arguments{
\item{START}{character. An adress of starting point.}

\item{STOP}{character. An adresss of destination.}
}
\value{
car_costs returns a cost of car travel pre traveler
}
\description{
Calculate total car travel cost
}
\examples{
car_travel_cost(START="plac Defilad 1, 00-901 Warszawa Polska",
STOP="Stanisława Kostki Potockiego 10/16, Warszawa Polska",
combustion=8, fuel_price=4.5, travelers_number = 1,
insurence_fee=1000, parking_time=8)
}
