% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{car_travel_cost}
\alias{car_travel_cost}
\title{Calculates cost for a given travel by car}
\usage{
car_travel_cost(start_point, finish_point, fuel_consumption, fuel_price,
  travelers_number = 1)
}
\arguments{
\item{start_point}{character. An address of start_pointing point.}

\item{finish_point}{character. An address of destination.}

\item{fuel_consumption}{numeric. Fuel consumption expressed as liters per 100 kilometers (L/100 km)}

\item{fuel_price}{numeric. Fuel price in PLN}

\item{travelers_number}{numeric. Number of travelers. Default value is 1.}
}
\value{
car_travel_cost returns a list containing following components:
  \item{distance}{distance of travel in km}
  \item{fuel_amount}{Amount of fuel in liters}
  \item{fuel_cost}{Amount of fuel in PLN}
  \item{fuel_cost_per_traveler}{Cost of fuel per traveler}
}
\description{
car_travel_cost calculates distans, amount of petrol, cost of petrol, cost of petrol for a given travel by car
}
\examples{
car_travel_cost(start_point="plac Defilad 1, 00-901 Warszawa Polska",
finish_point="plac Zamkowy 4 Warszawa Polska",
fuel_consumption=8, fuel_price=4.5, travelers_number = 1)
}
